refactor(toml): Pull out profile name validation
It now lives with other name validation logic.
