+++ src/cargo/util/command_prelude.rs
@@ -6,9 +6,8 @@ use crate::ops::{CompileFilter, CompileOptions, NewOptions, Packages, VersionCon
 use crate::util::important_paths::find_root_manifest_for_wd;
 use crate::util::interning::InternedString;
 use crate::util::is_rustup;
-use crate::util::restricted_names::is_glob_pattern;
-use crate::util::toml::schema::{StringOrVec, TomlProfile};
-use crate::util::validate_package_name;
+use crate::util::restricted_names;
+use crate::util::toml::schema::StringOrVec;
 use crate::util::{
     print_available_benches, print_available_binaries, print_available_examples,
     print_available_packages, print_available_tests,
@@ -607,7 +606,7 @@ Run `{cmd}` to see possible targets."
                 bail!("profile `doc` is reserved and not allowed to be explicitly specified")
             }
             (_, _, Some(name)) => {
-                TomlProfile::validate_name(name)?;
+                restricted_names::validate_profile_name(name)?;
                 name
             }
         };
@@ -801,7 +800,7 @@ Run `{cmd}` to see possible targets."
     ) -> CargoResult<CompileOptions> {
         let mut compile_opts = self.compile_options(config, mode, workspace, profile_checking)?;
         let spec = self._values_of("package");
-        if spec.iter().any(is_glob_pattern) {
+        if spec.iter().any(restricted_names::is_glob_pattern) {
             anyhow::bail!("Glob patterns on package selection are not supported.")
         }
         compile_opts.spec = Packages::Packages(spec);
@@ -835,7 +834,7 @@ Run `{cmd}` to see possible targets."
             (None, None) => config.default_registry()?.map(RegistryOrIndex::Registry),
             (None, Some(i)) => Some(RegistryOrIndex::Index(i.into_url()?)),
             (Some(r), None) => {
-                validate_package_name(r, "registry name", "")?;
+                restricted_names::validate_package_name(r, "registry name", "")?;
                 Some(RegistryOrIndex::Registry(r.to_string()))
             }
             (Some(_), Some(_)) => {
@@ -850,7 +849,7 @@ Run `{cmd}` to see possible targets."
         match self._value_of("registry").map(|s| s.to_string()) {
             None => config.default_registry(),
             Some(registry) => {
-                validate_package_name(&registry, "registry name", "")?;
+                restricted_names::validate_package_name(&registry, "registry name", "")?;
                 Ok(Some(registry))
             }
         }
