+++ src/cargo/util/toml/mod.rs
@@ -23,6 +23,7 @@ use crate::core::{GitReference, PackageIdSpec, SourceId, WorkspaceConfig, Worksp
 use crate::sources::{CRATES_IO_INDEX, CRATES_IO_REGISTRY};
 use crate::util::errors::{CargoResult, ManifestError};
 use crate::util::interning::InternedString;
+use crate::util::restricted_names;
 use crate::util::{
     self, config::ConfigRelativePath, validate_package_name, Config, IntoUrl, OptVersionReq,
     RustVersion,
@@ -2122,7 +2123,7 @@ impl schema::TomlProfile {
         }
 
         // Profile name validation
-        Self::validate_name(name)?;
+        restricted_names::validate_profile_name(name)?;
 
         if let Some(dir_name) = &self.dir_name {
             // This is disabled for now, as we would like to stabilize named
@@ -2175,85 +2176,6 @@ impl schema::TomlProfile {
                     (`\"thin\"`/`\"fat\"`/`\"off\"`) or omitted.",
                 );
             }
-        }
-
-        Ok(())
-    }
-
-    /// Validate dir-names and profile names according to RFC 2678.
-    pub fn validate_name(name: &str) -> CargoResult<()> {
-        if let Some(ch) = name
-            .chars()
-            .find(|ch| !ch.is_alphanumeric() && *ch != '_' && *ch != '-')
-        {
-            bail!(
-                "invalid character `{}` in profile name `{}`\n\
-                Allowed characters are letters, numbers, underscore, and hyphen.",
-                ch,
-                name
-            );
-        }
-
-        const SEE_DOCS: &str = "See https://doc.rust-lang.org/cargo/reference/profiles.html \
-            for more on configuring profiles.";
-
-        let lower_name = name.to_lowercase();
-        if lower_name == "debug" {
-            bail!(
-                "profile name `{}` is reserved\n\
-                 To configure the default development profile, use the name `dev` \
-                 as in [profile.dev]\n\
-                {}",
-                name,
-                SEE_DOCS
-            );
-        }
-        if lower_name == "build-override" {
-            bail!(
-                "profile name `{}` is reserved\n\
-                 To configure build dependency settings, use [profile.dev.build-override] \
-                 and [profile.release.build-override]\n\
-                 {}",
-                name,
-                SEE_DOCS
-            );
-        }
-
-        // These are some arbitrary reservations. We have no plans to use
-        // these, but it seems safer to reserve a few just in case we want to
-        // add more built-in profiles in the future. We can also uses special
-        // syntax like cargo:foo if needed. But it is unlikely these will ever
-        // be used.
-        if matches!(
-            lower_name.as_str(),
-            "build"
-                | "check"
-                | "clean"
-                | "config"
-                | "fetch"
-                | "fix"
-                | "install"
-                | "metadata"
-                | "package"
-                | "publish"
-                | "report"
-                | "root"
-                | "run"
-                | "rust"
-                | "rustc"
-                | "rustdoc"
-                | "target"
-                | "tmp"
-                | "uninstall"
-        ) || lower_name.starts_with("cargo")
-        {
-            bail!(
-                "profile name `{}` is reserved\n\
-                 Please choose a different name.\n\
-                 {}",
-                name,
-                SEE_DOCS
-            );
         }
 
         Ok(())
